move Keypair to its own file
